Adds initial test
