--- Java/Motorbike.java
@@ -1,7 +1,0 @@
-
-public class Motorbike implements Vehicle {
-  @Override
-  public String getType() {
-    return "Motorbike";
-  }
-}
