+++ Java/calculator/Car.java
@@ -1,0 +1,8 @@
+package calculator;
+
+public class Car implements Vehicle {
+  @Override
+  public String getType() {
+    return "calculator.Car";
+  }
+}
