+++ Java/calculator/Motorbike.java
@@ -1,0 +1,8 @@
+package calculator;
+
+public class Motorbike implements Vehicle {
+  @Override
+  public String getType() {
+    return "calculator.Motorbike";
+  }
+}
